+++ src/main/kotlin/model/Constants.kt
@@ -18,4 +18,6 @@ object Constants {
     const val DEFAULT_ZOOM_SCALE_COEF = .001f
 
     const val SEMI_BLACK = 0x50000000
+
+    const val SHOW_LABEL_MOD = 3
 }
