+++ src/main/kotlin/model/neo4j/EdgeEntity.kt
@@ -1,6 +1,11 @@
 package model.neo4j
 
-import org.neo4j.ogm.annotation.*
+import org.neo4j.ogm.annotation.EndNode
+import org.neo4j.ogm.annotation.GeneratedValue
+import org.neo4j.ogm.annotation.Id
+import org.neo4j.ogm.annotation.RelationshipEntity
+import org.neo4j.ogm.annotation.StartNode
+
 
 @RelationshipEntity(type = "ADJACENT TO")
 open class EdgeEntity @JvmOverloads constructor(
