+++ src/main/kotlin/model/neo4j/VertexEntity.kt
