+++ src/test/kotlin/model/MainScreenViewModelTest.kt
@@ -1,15 +1,13 @@
-package viewmodel
+package model
 
 import androidx.compose.ui.test.ExperimentalTestApi
 import androidx.compose.ui.test.assertIsNotDisplayed
 import androidx.compose.ui.test.onNodeWithTag
 import androidx.compose.ui.test.runComposeUiTest
 import kotlin.test.Test
-
-
-import model.DirectedGraph
 import space.kscience.kmath.operations.IntRing
 import view.MainScreenView
+import viewmodel.MainScreenViewModel
 
 class MainScreenViewModelTest {
