+++ src/test/kotlin/model/UndirectedGraphTest.kt
@@ -4,7 +4,8 @@ import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
-import org.junit.jupiter.api.Assertions.*
+import kotlin.test.assertEquals
+import kotlin.test.assertTrue
 import space.kscience.kmath.operations.IntRing
 
 @DisplayName("Tests for UndirectedGraph")
